Code Freezer
Patch relies on various modules for its internal functionality.  If the
test author patches one of these modules then Patch breaks.

The Code Freezer is a construct that allows Patch to intercept requests
to mock modules that are necessary for the proper functioning of Patch.

When the test author attempts to patch a "freezable" module, Patch will
make a frozen copy of that module and use that instead of the patched
module.  This allows the test author to patch any module without worry
that they will break Patch's functionality.

If a test author does not patch an internally critical module than the
normal module will be used and no penalty is paid by the test.