Bugfix for Assertion Reporting Race Conditions.
The assertion functions were susceptable to a race condition that was
technically correct but confusing for the test author.

Race Condition Description:

1. Assertion function checks the history for a matching call, none is
  found.
2. Assertion function pulls the history to format an error message.

Between 1 and 2 if a matching call arrived then the assertion would
fail but print out a message with a matching call.

This race condition has been defeated by only pulling the history once
and using it for both checking for the existence of a call and
formatting an error message.

In addition a bunch of code that was littering Patch.Mock has been
moved to a more appropriate location by introducing the concept of a
"Tagged History" (Patch.Mock.History.Tagged)

Tagged Histories have the same entries as a History.  One generates a
Tagged History from a History and some matching criteria.  Ever entry
in the Tagged History is the entry from the History tagged with a
boolean that indicates whether or not it matched the construction
criteria.