Fix a warning in Elixir 1.16
`1..-2` has an implicit `-1` step which is unused by String.slice but which generates a warning in Elixir 1.16

Elixir 1.16 suggests replacing this with `1..-2//1` but this syntax was introduced in 1.12 and would break any user of the library on older versions of Elixir.

This bugfix switches to `String.slice/3` which is available on all supported versions.