Move where mock execution happens
Previously, mocks were executed in the Patch.Mock.Server context.
With this change, mocks are executed in the callers context.

When providing a mock function it is a reasonable expectation that any function that cares about the current process would get the caller's process and not the Patch.Mock.Server.